Stop indicating non standard OIDs as mechanisms
We support interoperating on the wire with client using the pre-RFC
oid and the Microsoft wrong OIDs, we can also accept them from old
applications hardcoded to use those values. However we should not
advertize them at new applications that are asking the library what
mechanism are supported. The mechanism support is just the Kerberos
mechanism as per RFC.

Signed-off-by: Simo Sorce <simo@redhat.com>